requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
discord.py==1.5.1
dnspython==2.1.0
dpath==2.0.1
Flask==1.1.2
geojson==2.5.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.2
PyNaCl==1.3.0
pyowm==3.1.1
PySocks==1.7.1
pytz==2020.5
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7
selenium==3.141.0
six==1.15.0
soupsieve==2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.5.1
youtube-dl==2020.11.26
